HBASE-9696 Master recovery ignores online merge znode
git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1531435 13f79535-47bb-0310-9956-ffa450edef68

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/RegionMergeTransaction.java
@@ -18,6 +18,10 @@
  */
 package org.apache.hadoop.hbase.regionserver;
 
+import static org.apache.hadoop.hbase.executor.EventType.RS_ZK_REGION_MERGED;
+import static org.apache.hadoop.hbase.executor.EventType.RS_ZK_REGION_MERGING;
+import static org.apache.hadoop.hbase.executor.EventType.RS_ZK_REQUEST_REGION_MERGE;
+
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
@@ -45,6 +49,7 @@ import org.apache.hadoop.hbase.zookeeper.ZKUtil;
 import org.apache.hadoop.hbase.zookeeper.ZooKeeperWatcher;
 import org.apache.zookeeper.KeeperException;
 import org.apache.zookeeper.KeeperException.NodeExistsException;
+import org.apache.zookeeper.data.Stat;
 
 /**
  * Executes region merge as a "transaction". It is similar with
@@ -261,25 +266,16 @@ public class RegionMergeTransaction {
         createNodeMerging(server.getZooKeeper(), this.mergedRegionInfo,
           server.getServerName(), region_a.getRegionInfo(), region_b.getRegionInfo());
       } catch (KeeperException e) {
-        throw new IOException("Failed creating MERGING znode on "
+        throw new IOException("Failed creating PENDING_MERGE znode on "
             + this.mergedRegionInfo.getRegionNameAsString(), e);
       }
     }
     this.journal.add(JournalEntry.SET_MERGING_IN_ZK);
     if (server != null && server.getZooKeeper() != null) {
-      try {
-        // Transition node from MERGING to MERGING after creating the merge
-        // node. Master will get the callback for node change only if the
-        // transition is successful.
-        // Note that if the transition fails then the rollback will delete the
-        // created znode as the journal entry SET_MERGING_IN_ZK is added.
-        this.znodeVersion = transitionNodeMerging(server.getZooKeeper(),
-            this.mergedRegionInfo, server.getServerName(), -1,
-            region_a.getRegionInfo(), region_b.getRegionInfo());
-      } catch (KeeperException e) {
-        throw new IOException("Failed setting MERGING znode on "
-            + this.mergedRegionInfo.getRegionNameAsString(), e);
-      }
+      // After creating the merge node, wait for master to transition it
+      // from PENDING_MERGE to MERGING so that we can move on. We want master
+      // knows about it and won't transition any region which is merging.
+      znodeVersion = getZKNode(server, services);
     }
 
     this.region_a.getRegionFileSystem().createMergesDir();
@@ -303,9 +299,10 @@ public class RegionMergeTransaction {
       try {
         // Do one more check on the merging znode (before it is too late) in case
         // any merging region is moved somehow. If so, the znode transition will fail.
-        this.znodeVersion = transitionNodeMerging(server.getZooKeeper(),
-            this.mergedRegionInfo, server.getServerName(), this.znodeVersion,
-            region_a.getRegionInfo(), region_b.getRegionInfo());
+        this.znodeVersion = transitionMergingNode(server.getZooKeeper(),
+          this.mergedRegionInfo, region_a.getRegionInfo(), region_b.getRegionInfo(),
+          server.getServerName(), this.znodeVersion,
+          RS_ZK_REGION_MERGING, RS_ZK_REGION_MERGING);
       } catch (KeeperException e) {
         throw new IOException("Failed setting MERGING znode on "
             + this.mergedRegionInfo.getRegionNameAsString(), e);
@@ -489,9 +486,10 @@ public class RegionMergeTransaction {
 
     // Tell master about merge by updating zk. If we fail, abort.
     try {
-      this.znodeVersion = transitionNodeMerge(server.getZooKeeper(),
-          this.mergedRegionInfo, region_a.getRegionInfo(),
-          region_b.getRegionInfo(), server.getServerName(), this.znodeVersion);
+      this.znodeVersion = transitionMergingNode(server.getZooKeeper(),
+        this.mergedRegionInfo, region_a.getRegionInfo(),
+        region_b.getRegionInfo(), server.getServerName(), this.znodeVersion,
+        RS_ZK_REGION_MERGING, RS_ZK_REGION_MERGED);
 
       long startTime = EnvironmentEdgeManager.currentTimeMillis();
       int spins = 0;
@@ -506,9 +504,10 @@ public class RegionMergeTransaction {
         }
         Thread.sleep(100);
         // When this returns -1 it means the znode doesn't exist
-        this.znodeVersion = tickleNodeMerge(server.getZooKeeper(),
-            this.mergedRegionInfo, region_a.getRegionInfo(),
-            region_b.getRegionInfo(), server.getServerName(), this.znodeVersion);
+        this.znodeVersion = transitionMergingNode(server.getZooKeeper(),
+          this.mergedRegionInfo, region_a.getRegionInfo(),
+          region_b.getRegionInfo(), server.getServerName(), this.znodeVersion,
+          RS_ZK_REGION_MERGED, RS_ZK_REGION_MERGED);
         spins++;
       } while (this.znodeVersion != -1 && !server.isStopped()
           && !services.isStopping());
@@ -520,10 +519,81 @@ public class RegionMergeTransaction {
           + mergedRegionInfo.getEncodedName(), e);
     }
 
-
     // Leaving here, the mergedir with its dross will be in place but since the
     // merge was successful, just leave it; it'll be cleaned when region_a is
     // cleaned up by CatalogJanitor on master
+  }
+
+  /**
+   * Wait for the merging node to be transitioned from pending_merge
+   * to merging by master.  That's how we are sure master has processed
+   * the event and is good with us to move on. If we don't get any update,
+   * we periodically transition the node so that master gets the callback.
+   * If the node is removed or is not in pending_merge state any more,
+   * we abort the merge.
+   */
+  private int getZKNode(final Server server,
+      final RegionServerServices services) throws IOException {
+    // Wait for the master to process the pending_merge.
+    try {
+      int spins = 0;
+      Stat stat = new Stat();
+      ZooKeeperWatcher zkw = server.getZooKeeper();
+      ServerName expectedServer = server.getServerName();
+      String node = mergedRegionInfo.getEncodedName();
+      while (!(server.isStopped() || services.isStopping())) {
+        if (spins % 5 == 0) {
+          LOG.debug("Still waiting for master to process "
+            + "the pending_merge for " + node);
+          transitionMergingNode(zkw, mergedRegionInfo, region_a.getRegionInfo(),
+            region_b.getRegionInfo(), expectedServer, -1, RS_ZK_REQUEST_REGION_MERGE,
+            RS_ZK_REQUEST_REGION_MERGE);
+        }
+        Thread.sleep(100);
+        spins++;
+        byte [] data = ZKAssign.getDataNoWatch(zkw, node, stat);
+        if (data == null) {
+          throw new IOException("Data is null, merging node "
+            + node + " no longer exists");
+        }
+        RegionTransition rt = RegionTransition.parseFrom(data);
+        EventType et = rt.getEventType();
+        if (et == RS_ZK_REGION_MERGING) {
+          ServerName serverName = rt.getServerName();
+          if (!serverName.equals(expectedServer)) {
+            throw new IOException("Merging node " + node + " is for "
+              + serverName + ", not us " + expectedServer);
+          }
+          byte [] payloadOfMerging = rt.getPayload();
+          List<HRegionInfo> mergingRegions = HRegionInfo.parseDelimitedFrom(
+            payloadOfMerging, 0, payloadOfMerging.length);
+          assert mergingRegions.size() == 3;
+          HRegionInfo a = mergingRegions.get(1);
+          HRegionInfo b = mergingRegions.get(2);
+          HRegionInfo hri_a = region_a.getRegionInfo();
+          HRegionInfo hri_b = region_b.getRegionInfo();
+          if (!(hri_a.equals(a) && hri_b.equals(b))) {
+            throw new IOException("Merging node " + node + " is for " + a + ", "
+              + b + ", not expected regions: " + hri_a + ", " + hri_b);
+          }
+          // Master has processed it.
+          return stat.getVersion();
+        }
+        if (et != RS_ZK_REQUEST_REGION_MERGE) {
+          throw new IOException("Merging node " + node
+            + " moved out of merging to " + et);
+        }
+      }
+      // Server is stopping/stopped
+      throw new IOException("Server is "
+        + (services.isStopping() ? "stopping" : "stopped"));
+    } catch (Exception e) {
+      if (e instanceof InterruptedException) {
+        Thread.currentThread().interrupt();
+      }
+      throw new IOException("Failed getting MERGING znode on "
+        + mergedRegionInfo.getRegionNameAsString(), e);
+    }
   }
 
   /**
@@ -566,6 +636,7 @@ public class RegionMergeTransaction {
    *         of no return and so now need to abort the server to minimize
    *         damage.
    */
+  @SuppressWarnings("deprecation")
   public boolean rollback(final Server server,
       final RegionServerServices services) throws IOException {
     assert this.mergedRegionInfo != null;
@@ -653,20 +724,22 @@ public class RegionMergeTransaction {
   private static void cleanZK(final Server server, final HRegionInfo hri) {
     try {
       // Only delete if its in expected state; could have been hijacked.
-      ZKAssign.deleteNode(server.getZooKeeper(), hri.getEncodedName(),
-          EventType.RS_ZK_REGION_MERGING);
+      if (!ZKAssign.deleteNode(server.getZooKeeper(), hri.getEncodedName(),
+          RS_ZK_REQUEST_REGION_MERGE)) {
+        ZKAssign.deleteNode(server.getZooKeeper(), hri.getEncodedName(),
+          RS_ZK_REGION_MERGING);
+      }
     } catch (KeeperException.NoNodeException e) {
       LOG.warn("Failed cleanup zk node of " + hri.getRegionNameAsString(), e);
     } catch (KeeperException e) {
       server.abort("Failed cleanup zk node of " + hri.getRegionNameAsString(),e);
     }
-
   }
 
   /**
-   * Creates a new ephemeral node in the MERGING state for the merged region.
+   * Creates a new ephemeral node in the PENDING_MERGE state for the merged region.
    * Create it ephemeral in case regionserver dies mid-merge.
-   * 
+   *
    * <p>
    * Does not transition nodes from other states. If a node already exists for
    * this region, a {@link NodeExistsException} will be thrown.
@@ -674,32 +747,27 @@ public class RegionMergeTransaction {
    * @param zkw zk reference
    * @param region region to be created as offline
    * @param serverName server event originates from
-   * @return Version of znode created.
    * @throws KeeperException
    * @throws IOException
    */
-  int createNodeMerging(final ZooKeeperWatcher zkw, final HRegionInfo region,
+  public static void createNodeMerging(final ZooKeeperWatcher zkw, final HRegionInfo region,
       final ServerName serverName, final HRegionInfo a,
       final HRegionInfo b) throws KeeperException, IOException {
     LOG.debug(zkw.prefix("Creating ephemeral node for "
-        + region.getEncodedName() + " in MERGING state"));
-    byte [] payload = HRegionInfo.toDelimitedByteArray(a, b);
+      + region.getEncodedName() + " in PENDING_MERGE state"));
+    byte [] payload = HRegionInfo.toDelimitedByteArray(region, a, b);
     RegionTransition rt = RegionTransition.createRegionTransition(
-        EventType.RS_ZK_REGION_MERGING, region.getRegionName(), serverName, payload);
+      RS_ZK_REQUEST_REGION_MERGE, region.getRegionName(), serverName, payload);
     String node = ZKAssign.getNodeName(zkw, region.getEncodedName());
     if (!ZKUtil.createEphemeralNodeAndWatch(zkw, node, rt.toByteArray())) {
       throw new IOException("Failed create of ephemeral " + node);
     }
-    // Transition node from MERGING to MERGING and pick up version so we
-    // can be sure this znode is ours; version is needed deleting.
-    return transitionNodeMerging(zkw, region, serverName, -1, a, b);
   }
 
   /**
-   * Transitions an existing node for the specified region which is currently in
-   * the MERGING state to be in the MERGE state. Converts the ephemeral MERGING
-   * znode to an ephemeral MERGE node. Master cleans up MERGE znode when it
-   * reads it (or if we crash, zk will clean it up).
+   * Transitions an existing ephemeral node for the specified region which is
+   * currently in the begin state to be in the end state. Master cleans up the
+   * final MERGE znode when it reads it (or if we crash, zk will clean it up).
    *
    * <p>
    * Does not transition nodes from other states. If for some reason the node
@@ -710,19 +778,18 @@ public class RegionMergeTransaction {
    * This method can fail and return false for three different reasons:
    * <ul>
    * <li>Node for this region does not exist</li>
-   * <li>Node for this region is not in MERGING state</li>
-   * <li>After verifying MERGING state, update fails because of wrong version
+   * <li>Node for this region is not in the begin state</li>
+   * <li>After verifying the begin state, update fails because of wrong version
    * (this should never actually happen since an RS only does this transition
-   * following a transition to MERGING. if two RS are conflicting, one would
-   * fail the original transition to MERGING and not this transition)</li>
+   * following a transition to the begin state. If two RS are conflicting, one would
+   * fail the original transition to the begin state and not this transition)</li>
    * </ul>
    *
    * <p>
    * Does not set any watches.
    *
    * <p>
-   * This method should only be used by a RegionServer when completing the open
-   * of merged region.
+   * This method should only be used by a RegionServer when merging two regions.
    *
    * @param zkw zk reference
    * @param merged region to be transitioned to opened
@@ -730,45 +797,19 @@ public class RegionMergeTransaction {
    * @param b merging region B
    * @param serverName server event originates from
    * @param znodeVersion expected version of data before modification
+   * @param beginState the expected current state the znode should be
+   * @param endState the state to be transition to
    * @return version of node after transition, -1 if unsuccessful transition
    * @throws KeeperException if unexpected zookeeper exception
    * @throws IOException
    */
-  private static int transitionNodeMerge(ZooKeeperWatcher zkw,
+  public static int transitionMergingNode(ZooKeeperWatcher zkw,
       HRegionInfo merged, HRegionInfo a, HRegionInfo b, ServerName serverName,
-      final int znodeVersion) throws KeeperException, IOException {
+      final int znodeVersion, final EventType beginState,
+      final EventType endState) throws KeeperException, IOException {
     byte[] payload = HRegionInfo.toDelimitedByteArray(merged, a, b);
     return ZKAssign.transitionNode(zkw, merged, serverName,
-        EventType.RS_ZK_REGION_MERGING, EventType.RS_ZK_REGION_MERGED,
-        znodeVersion, payload);
-  }
-
-  /**
-   *
-   * @param zkw zk reference
-   * @param parent region to be transitioned to merging
-   * @param serverName server event originates from
-   * @param version znode version
-   * @return version of node after transition, -1 if unsuccessful transition
-   * @throws KeeperException
-   * @throws IOException
-   */
-  int transitionNodeMerging(final ZooKeeperWatcher zkw,
-      final HRegionInfo parent, final ServerName serverName, final int version,
-      final HRegionInfo a, final HRegionInfo b) throws KeeperException, IOException {
-    byte[] payload = HRegionInfo.toDelimitedByteArray(a, b);
-    return ZKAssign.transitionNode(zkw, parent, serverName,
-            EventType.RS_ZK_REGION_MERGING, EventType.RS_ZK_REGION_MERGING,
-        version, payload);
-  }
-
-  private static int tickleNodeMerge(ZooKeeperWatcher zkw, HRegionInfo merged,
-      HRegionInfo a, HRegionInfo b, ServerName serverName,
-      final int znodeVersion) throws KeeperException, IOException {
-    byte[] payload = HRegionInfo.toDelimitedByteArray(a, b);
-    return ZKAssign.transitionNode(zkw, merged, serverName,
-        EventType.RS_ZK_REGION_MERGED, EventType.RS_ZK_REGION_MERGED,
-        znodeVersion, payload);
+      beginState, endState, znodeVersion, payload);
   }
 
   /**
